Add build support for Arduino Teensy
The Arduino Teensy has a different processor architecture, and
therefore a different build chain, than standard Arduinos. Building
and uploading are still possible through the Arduino IDE by patching
it and installing the teensy build chain. This patch adds parsing and
handling of extra options in the teensy's boards.txt, which should
allow it to read all necessary configuration to build on all Teensy
versions.

With this patch, the following command will build for teensy3 with a
96 MHz clock, an en-US keyboard format, and serial USB mode:

 $ ino build -m teensy3 --menu='keys:en-us,speed:96,usb:serial'

** This patch supports building only. Uploading will come later.

*** For some reason, this patch does not work with teensy1. I don't
know why right now, but I will investigate.